rename the fuse op, test=allcase
