+++ paddle/fluid/operators/fused_softmax_mask_upper_triangle_op.h
